fix: fixed  deadlocks; fixed usleeps;

+++ philo/srcs/utils/ft_get_time.c
@@ -6,13 +6,13 @@
 /*   By: epakdama <epakdama@student.42istanbul.c    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/08/11 10:54:58 by epakdama          #+#    #+#             */
-/*   Updated: 2025/08/11 10:57:40 by epakdama         ###   ########.fr       */
+/*   Updated: 2025/08/11 12:23:37 by epakdama         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_philo.h"
 
-size_t	ft_get_time(t_data *data)
+long long	ft_get_time(t_data *data)
 {
 	struct timeval	tv;
